requirements.txt
Flask==1.1.1
numpy==1.19.1
gunicorn==19.9.0
Werkzeug==1.0.0
tensorflow-cpu==2.5.0
pillow>=6.2.0